+++ tensorflow/lite/micro/memory_planner/linear_memory_planner.h
@@ -27,17 +27,17 @@ class LinearMemoryPlanner : public MemoryPlanner {
   LinearMemoryPlanner();
   ~LinearMemoryPlanner() override;
 
-  TfLiteStatus AddBuffer(tflite::ErrorReporter* error_reporter, int size,
+  TfLiteStatus AddBuffer(tflite::ErrorReporter* error_reporter, size_t size,
                          int first_time_used, int last_time_used) override;
 
-  int GetMaximumMemorySize() override;
+  size_t GetMaximumMemorySize() override;
   int GetBufferCount() override;
   TfLiteStatus GetOffsetForBuffer(tflite::ErrorReporter* error_reporter,
-                                  int buffer_index, int* offset) override;
+                                  int buffer_index, size_t* offset) override;
 
  private:
   static constexpr int kMaxBufferCount = 1024;
-  int buffer_offsets_[kMaxBufferCount];
+  size_t buffer_offsets_[kMaxBufferCount];
   int current_buffer_count_;
   int next_free_offset_;
 };
